Move drawable chat related classes to better namespace
