+++ osu.Game/osu.Game.csproj
@@ -304,8 +304,8 @@
     <Compile Include="Online\API\Requests\GetMessagesRequest.cs" />
     <Compile Include="Online\API\Requests\ListChannelsRequest.cs" />
     <Compile Include="Online\Chat\Channel.cs" />
-    <Compile Include="Online\Chat\Drawables\DrawableChannel.cs" />
-    <Compile Include="Online\Chat\Drawables\ChatLine.cs" />
+    <Compile Include="Overlays\Chat\DrawableChannel.cs" />
+    <Compile Include="Overlays\Chat\ChatLine.cs" />
     <Compile Include="Online\Chat\Message.cs" />
     <Compile Include="OsuGame.cs" />
     <Compile Include="OsuGameBase.cs" />
